Disable plugins with unmetdependencies, in admin home and plugins.php admin page. See #1842, fixed partially (need to see how to cope with themes now)

+++ admin/index.php
@@ -28,6 +28,10 @@ if (!empty($_GET['default_blog'])) {
 
 dcPage::check('usage,contentadmin');
 
+if ($core->plugins->disableDepModules($core->adminurl->get('admin.home',array()))) {
+	exit;
+}
+
 # Logout
 if (!empty($_GET['logout'])) {
 	$core->session->destroy();
